DAQ reader: First draft of init, exit, read data functions
The select system call is used to determine when the metadata device
can deliver data.
The data can then be read through DMA is is sent to wireshark.
The function to send data to wireshark is noch fully implemented yet.

The p40_daq_* functions are used in order to start and stop data
generation and retrieve the pointer to the buffer, which contains
the data in DMA memory region.

For now, mock values for device numbers are used.