Replaced defines by cfg fields
Before, the defines, which determined the field width, were directly
passed to proto_tree_add_bits_item. Now, a config struct, which is
initialized on startup is used for that.

Therefore, in the future, the configuration can be read from a config
file and set on startup without the need to recompile the plugin.